refactor: separate pure logic from DOM utilities into utils-logic.js
