+++ js/utils/utils.js
@@ -1,4 +1,4 @@
-// Copyright (c) 2014-21 Walter Bender
+// Copyright (c) 2014-2024 Walter Bender
 //
 // This program is free software; you can redistribute it and/or
 // modify it under the terms of the The GNU Affero General Public
@@ -14,36 +14,30 @@
 
    PALETTEICONS, PALETTEFILLCOLORS, PALETTESTROKECOLORS,
    PALETTEHIGHLIGHTCOLORS, HIGHLIGHTSTROKECOLORS, MULTIPALETTES,
-   platformColor, base64Encode, i18next, createjs
-*/
-
-/*
-    Global locations
-    - js/artwork.js
-        PALETTEICONS, PALETTEFILLCOLORS, PALETTESTROKECOLORS, PALETTEHIGHLIGHTCOLORS,
-        HIGHLIGHTSTROKECOLORS
-    - js/turtledefs.js
-        MULTIPALETTES
-    - js/utils/platformstyle.js
-        platformColor
+   platformColor, base64Encode, i18next, createjs, Howler, jQuery,
+   ActivityContext, TITLESTRING, _THIS_IS_MUSIC_BLOCKS_, _THIS_IS_TURTLE_BLOCKS_
 */
 
 /* exported
 
    canvasPixelRatio, changeImage, closeBlkWidgets, closeWidgets,
-   deepClone, delayExecution, displayMsg, doBrowserCheck, docByClass, docByName,
+   delayExecution, displayMsg, doBrowserCheck, docByClass, docByName,
    docBySelector, docByTagName, doPublish, doStopVideoCam, doSVG,
-   doUseCamera, fileBasename, fileExt, format, getTextWidth, hex2rgb,
-   hexToRGB, hideDOMLabel, httpGet, httpPost, HttpRequest,
-   importMembers, isSVGEmpty, last, mixedNumber, nearestBeat,
-   oneHundredToFraction, prepareMacroExports, preparePluginExports,
-   processMacroData, processRawPluginData, rationalSum, rgbToHex,
-   safeSVG, safeJSONParse, toFixed2, toTitleCase, windowHeight, windowWidth,
-    fnBrowserDetect, waitForReadiness, isSafeUrl, unescapeHTML
+   doUseCamera, format, getTextWidth, hideDOMLabel, httpGet, httpPost,
+   HttpRequest, importMembers, isSVGEmpty, prepareMacroExports,
+   preparePluginExports, processMacroData, processRawPluginData,
+   windowHeight, windowWidth, fnBrowserDetect, waitForReadiness
 */
 
 /**
+ * @fileoverview Browser-dependent utility functions for Music Blocks.
+ * These functions require a DOM environment (window, document, navigator).
+ * Pure logic functions have been moved to js/utils/utils-logic.js.
+ */
+
+/**
  * Changes the source of an image element from one SVG data URI to another.
+ * @requires browser
  * @function
  * @param {HTMLImageElement} imgElement - The image element to update.
  * @param {string} from - The base64-encoded SVG data URI to replace.
@@ -58,34 +52,9 @@ const changeImage = (imgElement, from, to) => {
 };
 
 /**
- * Safely parses a JSON string, wrapping the operation in a try/catch block
- * to prevent the application from crashing on malformed JSON payload data (e.g. from localStorage).
- *
- * @function
- * @param {string} data - The JSON string to parse
- * @param {*} fallback - The fallback value to return if JSON.parse throws an error. Defaults to null.
- * @returns {*} The successfully parsed Object/Array, or the fallback value upon failure.
- */
-const safeJSONParse = (data, fallback = null) => {
-    if (typeof data !== "string" || !data) return fallback;
-    try {
-        return JSON.parse(data);
-    } catch (e) {
-        console.warn("Failed to safely parse JSON:", e);
-        return fallback;
-    }
-};
-
-if (typeof module !== "undefined" && module.exports) {
-    module.exports.safeJSONParse = safeJSONParse;
-}
-if (typeof window !== "undefined") {
-    window.safeJSONParse = safeJSONParse;
-}
-
-/**
  * Enhanced _() method to handle case variations for translations
  * prioritize exact matches and preserve the case of the input text.
+ * @requires browser (localStorage, i18next)
  * @function
  * @param {string} text - The input text to be translated.
  * @returns {string} The translated text.
@@ -153,6 +122,7 @@ function _(text, options = {}) {
 
 /**
  * A string formatting function using placeholder substitution.
+ * @requires browser (calls _())
  * @function
  * @param {string} str - The template string with placeholders.
  * @param {Object} data - The data object containing values for substitution.
@@ -179,6 +149,7 @@ let format = (str, data) => {
 
 /**
  * Detects the current browser name.
+ * @requires browser (navigator)
  * @function
  * @returns {string} The name of the detected browser.
  */
@@ -204,6 +175,7 @@ function fnBrowserDetect() {
 
 /**
  * Returns the pixel ratio of the canvas for high-resolution displays.
+ * @requires browser (window, document)
  * @function
  * @returns {number} The canvas pixel ratio.
  */
@@ -222,6 +194,7 @@ function canvasPixelRatio() {
 
 /**
  * Returns the height of the window, accounting for Android-specific behavior.
+ * @requires browser (window, navigator)
  * @function
  * @returns {number} The window height.
  */
@@ -236,6 +209,7 @@ function windowHeight() {
 
 /**
  * Returns the width of the window, accounting for Android-specific behavior.
+ * @requires browser (window, navigator)
  * @function
  * @returns {number} The window width.
  */
@@ -250,9 +224,8 @@ function windowWidth() {
 
 /**
  * Performs an HTTP GET request to retrieve data from the server.
- * Uses async fetch to avoid blocking the UI during network requests.
+ * @requires browser (fetch, window.server)
  * @param {string|null} projectName - The name of the project (or null for the base URL).
- * @throws {Error} Throws an error if the HTTP status code is greater than 299.
  * @returns {Promise<string>} A promise that resolves to the response text from the server.
  */
 let httpGet = async projectName => {
@@ -273,7 +246,7 @@ let httpGet = async projectName => {
 
 /**
  * Performs an HTTP POST request to send data to the server.
- * Uses async fetch to avoid blocking the UI during network requests.
+ * @requires browser (fetch, window.server)
  * @param {string} projectName - The name of the project.
  * @param {string} data - The data to be sent in the POST request.
  * @returns {Promise<string>} A promise that resolves to the response text from the server.
@@ -296,6 +269,7 @@ let httpPost = async (projectName, data) => {
 
 /**
  * Constructor function for making an HTTP request.
+ * @requires browser (XMLHttpRequest, self.location)
  * @constructor
  * @param {string} url - The URL to make the HTTP request to.
  * @param {function} loadCallback - The callback function to handle the loaded response.
@@ -335,7 +309,7 @@ function HttpRequest(url, loadCallback, userCallback) {
 
 /**
  * Checks the browser type and version.
- * Sets properties in the jQuery.browser object based on the user agent.
+ * @requires browser (navigator, jQuery)
  * @function
  */
 function doBrowserCheck() {
@@ -375,23 +349,13 @@ function doBrowserCheck() {
 
 /**
  * Wait for critical dependencies to be ready before calling callback.
- * Uses polling with exponential backoff and maximum timeout.
- * This replaces the arbitrary 5-second delay for Firefox with actual readiness checks.
- *
+ * @requires browser (document, createjs, Howler, jQuery, requestAnimationFrame)
  * @param {Function} callback - The function to call when ready
- * @param {Object} options - Configuration options
- * @param {number} options.maxWait - Maximum wait time in ms (default: 10000)
- * @param {number} options.minWait - Minimum wait time in ms (default: 500)
- * @param {number} options.checkInterval - Initial check interval in ms (default: 100)
  */
 function waitForReadiness(callback, options = {}) {
     const { maxWait = 10000, minWait = 500, checkInterval = 100 } = options;
     const startTime = Date.now();
 
-    /**
-     * Check if critical dependencies and DOM elements are ready
-     * @returns {boolean} True if all critical dependencies are loaded
-     */
     const isReady = () => {
         // Check if critical JavaScript libraries are loaded
         const createjsLoaded = typeof createjs !== "undefined" && createjs.Stage;
@@ -407,36 +371,26 @@ function waitForReadiness(callback, options = {}) {
         return createjsLoaded && howlerLoaded && jqueryLoaded && domReady;
     };
 
-    /**
-     * Polling function that checks readiness and calls callback when ready
-     */
     const check = () => {
         const elapsed = Date.now() - startTime;
 
         if (elapsed >= minWait && isReady()) {
-            // Ready! Initialize the app
-
             console.log(`[Firefox] Initialized in ${elapsed}ms (readiness-based)`);
             callback();
         } else if (elapsed >= maxWait) {
-            // Timeout - initialize anyway as fallback
-
             console.warn(
                 `[Firefox] Initialization timed out after ${maxWait}ms, proceeding anyway`
             );
             callback();
         } else {
-            // Not ready yet, check again on next animation frame
             requestAnimationFrame(check);
         }
     };
 
-    // Start the readiness check loop
     requestAnimationFrame(check);
 }
 
 // Check for Internet Explorer
-
 window.addEventListener("load", () => {
     const userAgent = window.navigator.userAgent;
     // For IE 10 or older
@@ -488,8 +442,7 @@ window.addEventListener("load", () => {
 
 /**
  * Retrieves a collection of elements by class name.
- * @param {string} classname - The class name to search for.
- * @returns {HTMLCollectionOf<Element>} A collection of elements with the specified class name.
+ * @requires browser (document)
  */
 function docByClass(classname) {
     return document.getElementsByClassName(classname);
@@ -497,8 +450,7 @@ function docByClass(classname) {
 
 /**
  * Retrieves a collection of elements by tag name.
- * @param {string} tag - The tag name to search for.
- * @returns {NodeList} A collection of elements with the specified tag name.
+ * @requires browser (document)
  */
 function docByTagName(tag) {
     return document.getElementsByTagName(tag);
@@ -506,8 +458,7 @@ function docByTagName(tag) {
 
 /**
  * Retrieves an element by its ID.
- * @param {string} id - The ID of the element to retrieve.
- * @returns {HTMLElement|null} The element with the specified ID, or null if not found.
+ * @requires browser (document)
  */
 function docById(id) {
     return document.getElementById(id);
@@ -515,8 +466,7 @@ function docById(id) {
 
 /**
  * Retrieves a collection of elements by name.
- * @param {string} name - The name attribute value to search for.
- * @returns {NodeListOf<Element>} A collection of elements with the specified name attribute.
+ * @requires browser (document)
  */
 function docByName(name) {
     return document.getElementsByName(name);
@@ -524,46 +474,15 @@ function docByName(name) {
 
 /**
  * Retrieves the first element that matches a specified CSS selector.
- * @param {string} selector - A CSS selector string.
- * @returns {Element|null} The first element that matches the selector, or null if not found.
+ * @requires browser (document)
  */
 function docBySelector(selector) {
     return document.querySelector(selector);
 }
 
 /**
- * Returns the last element of an array.
- * @param {Array} myList - The array from which to get the last element.
- * @returns {*} The last element of the array, or null if the array is empty.
- */
-let last = myList => {
-    const i = myList.length;
-    if (i === 0) {
-        return null;
-    } else {
-        return myList[i - 1];
-    }
-};
-
-/**
- * Creates a deep clone of a value. Uses structuredClone when available
- * (modern browsers) for better performance, falling back to
- * JSON.parse(JSON.stringify()) for compatibility with test environments.
- * @param {*} value - The value to deep clone.
- * @returns {*} A deep clone of the value.
- */
-let deepClone = value => {
-    if (typeof structuredClone === "function") {
-        return structuredClone(value);
-    }
-    return JSON.parse(JSON.stringify(value));
-};
-
-/**
  * Gets the width of a text string given a specific font.
- * @param {string} text - The text string.
- * @param {string} font - The font style and size.
- * @returns {number} The width of the text in pixels.
+ * @requires browser (document)
  */
 let getTextWidth = (text, font) => {
     // re-use canvas object for better performance
@@ -576,13 +495,6 @@ let getTextWidth = (text, font) => {
 
 /**
  * Generates an SVG string representing the canvas, logo, and turtles.
- * @param {object} canvas - The canvas object.
- * @param {object} logo - The logo object.
- * @param {object} turtles - The turtles object.
- * @param {number} width - The width of the SVG.
- * @param {number} height - The height of the SVG.
- * @param {number} scale - The scaling factor.
- * @returns {string} The SVG string.
  */
 function doSVG(canvas, logo, turtles, width, height, scale) {
     // Aggregate SVG output from each turtle. If there is none, return an empty string.
@@ -615,8 +527,6 @@ function doSVG(canvas, logo, turtles, width, height, scale) {
 
 /**
  * Checks if the SVG output from turtles is empty.
- * @param {object} turtles - The turtles object.
- * @returns {boolean} True if all turtle SVG outputs are empty, false otherwise.
  */
 let isSVGEmpty = turtles => {
     for (const turtle in turtles.turtleList) {
@@ -629,135 +539,8 @@ let isSVGEmpty = turtles => {
 };
 
 /**
- * Gets the file extension from a file path.
- * @param {string} file - The file path or name.
- * @returns {string} The file extension.
- */
-let fileExt = file => {
-    if (file === null) {
-        return "";
-    }
-
-    const parts = file.split(".");
-    if (parts.length === 1 || (parts[0] === "" && parts.length === 2)) {
-        return "";
-    }
-
-    return parts.pop();
-};
-
-/**
- * Gets the basename of a file path (excluding the extension).
- * @param {string} file - The file path or name.
- * @returns {string} The basename.
- */
-let fileBasename = file => {
-    const parts = file.split(".");
-    if (parts.length === 1) {
-        return parts[0];
-    } else if (parts[0] === "" && parts.length === 2) {
-        return file;
-    } else {
-        parts.pop(); // throw away suffix
-        return parts.join(".");
-    }
-};
-
-/**
- * Converts the first character of a string to uppercase.
- * @param {string} str - The input string.
- * @returns {string} The string with the first character in uppercase.
- */
-function toTitleCase(str) {
-    if (typeof str !== "string") return;
-    if (str.length === 0) return "";
-    return str.charAt(0).toUpperCase() + str.slice(1);
-}
-
-if (typeof module !== "undefined" && module.exports) {
-    module.exports.toTitleCase = toTitleCase;
-}
-if (typeof window !== "undefined") {
-    window.toTitleCase = toTitleCase;
-}
-
-/**
- * Escapes HTML special characters to prevent XSS when injecting
- * user-provided values into HTML.
- * @param {string} str - The string to escape.
- * @returns {string} The escaped string safe for HTML insertion.
- */
-function escapeHTML(str) {
-    const escapeMap = {
-        "&": "&amp;",
-        "<": "&lt;",
-        ">": "&gt;",
-        '"': "&quot;",
-        "'": "&#039;"
-    };
-    return String(str).replace(/[&<>"']/g, char => escapeMap[char]);
-}
-
-if (typeof module !== "undefined" && module.exports) {
-    module.exports.escapeHTML = escapeHTML;
-}
-if (typeof window !== "undefined") {
-    window.escapeHTML = escapeHTML;
-}
-
-/**
- * Reverses HTML entity escaping produced by escapeHTML().
- * Used when loading project data that was escaped for safe HTML embedding.
- * @param {string} str - The HTML-escaped string to unescape.
- * @returns {string} The unescaped string with original characters restored.
- */
-function unescapeHTML(str) {
-    const unescapeMap = {
-        "&amp;": "&",
-        "&lt;": "<",
-        "&gt;": ">",
-        "&quot;": '"',
-        "&#039;": "'"
-    };
-    return String(str).replace(/&amp;|&lt;|&gt;|&quot;|&#039;/g, match => unescapeMap[match]);
-}
-
-if (typeof module !== "undefined" && module.exports) {
-    module.exports.unescapeHTML = unescapeHTML;
-}
-if (typeof window !== "undefined") {
-    window.unescapeHTML = unescapeHTML;
-}
-
-/**
- * Validates that a URL string uses a safe protocol (http or https).
- * Uses the URL API for robust parsing instead of fragile regex patterns.
- * This prevents open redirect attacks via javascript:, data:, vbscript:,
- * or other dangerous URI schemes.
- * @param {string} urlString - The URL string to validate.
- * @returns {boolean} True if the URL uses http: or https: protocol.
- */
-function isSafeUrl(urlString) {
-    try {
-        const parsed = new URL(urlString);
-        return parsed.protocol === "http:" || parsed.protocol === "https:";
-    } catch (e) {
-        return false;
-    }
-}
-
-if (typeof module !== "undefined" && module.exports) {
-    module.exports.isSafeUrl = isSafeUrl;
-}
-if (typeof window !== "undefined") {
-    window.isSafeUrl = isSafeUrl;
-}
-
-/**
  * Processes plugin data and updates the activity based on the provided JSON-encoded dictionary.
- * @param {object} activity - The activity object to update.
- * @param {string} pluginData - The JSON-encoded plugin data.
- * @returns {object|null} The processed plugin data object or null if parsing fails.
+ * @requires browser (confirm, localStorage, document)
  */
 const processPluginData = async (activity, pluginData, pluginSource) => {
     // Plugins are JSON-encoded dictionaries.
@@ -1294,10 +1077,6 @@ let preparePluginExports = (activity, obj) => {
 
 /**
  * Processes macro data, adds macros to the palette, and updates the macro dictionary.
- * @param {string} macroData - JSON-encoded dictionary containing macro data.
- * @param {object} palettes - The palettes object.
- * @param {object} blocks - The blocks object.
- * @param {object} macroDict - The macro dictionary to update.
  */
 let processMacroData = (macroData, palettes, blocks, macroDict) => {
     // Macros are stored in a JSON-encoded dictionary.
@@ -1353,13 +1132,7 @@ const CameraManager = {
 
 /**
  * Uses the camera to capture images or video frames and displays them on the turtle's canvas.
- * @param {Array} args - Arguments passed to the function.
- * @param {object} turtles - The turtles object.
- * @param {string} turtle - The name of the turtle.
- * @param {boolean} isVideo - Indicates whether to capture video frames.
- * @param {number} cameraID - The ID of the camera interval for video frames.
- * @param {function} setCameraID - Function to set the camera interval ID.
- * @param {function} errorMsg - Function to display error messages.
+ * @requires browser (navigator, video, canvas)
  */
 let doUseCamera = (args, turtles, turtle, isVideo, cameraID, setCameraID, errorMsg) => {
     const w = 320;
@@ -1451,8 +1224,7 @@ let doUseCamera = (args, turtles, turtle, isVideo, cameraID, setCameraID, errorM
 
 /**
  * Stops the camera and clears the camera interval.
- * @param {number} cameraID - The ID of the camera interval.
- * @param {function} setCameraID - Function to set the camera interval ID.
+ * @requires browser
  */
 function doStopVideoCam(cameraID, setCameraID) {
     if (cameraID !== null) {
@@ -1474,6 +1246,7 @@ function doStopVideoCam(cameraID, setCameraID) {
 
 /**
  * Hides certain DOM elements related to labels.
+ * @requires browser
  */
 function hideDOMLabel() {
     const textLabel = docById("textLabel");
@@ -1494,461 +1267,14 @@ function hideDOMLabel() {
 
 /**
  * Displays a message (currently unused).
- * @returns {undefined}
  */
 function displayMsg(/*blocks, text*/) {
-    /*
-    let msgContainer = blocks.msgText.parent;
-    msgContainer.visible = true;
-    blocks.msgText.text = text;
-    msgContainer.updateCache();
-    blocks.stage.setChildIndex(msgContainer, blocks.stage.getNumChildren() - 1);
-    */
     return;
 }
 
 /**
- * Escapes HTML entities in a given string to make it safe for SVG.
- * @param {string} label - The string to escape.
- * @returns {string} The escaped string.
- */
-function safeSVG(label) {
-    if (typeof label === "string") {
-        return label.replace(/&/g, "&amp;").replace(/</g, "&lt;").replace(/>/g, "&gt;");
-    } else {
-        return label;
-    }
-}
-
-/**
- * Formats a number to fixed two precision.
- * @param {number} d - The number to format.
- * @returns {string} The formatted number.
- */
-function toFixed2(d) {
-    // Return number as fixed 2 precision
-    if (typeof d === "number") {
-        const floor = Math.floor(d);
-        if (d !== floor) {
-            return d.toFixed(2).toString();
-        } else {
-            return d.toString();
-        }
-    } else {
-        return d;
-    }
-}
-
-/**
- * Converts a floating-point number to its approximate fractional representation.
- * @param {number} d - The input number.
- * @returns {Array} An array representing the fraction in the form [numerator, denominator].
- */
-function rationalToFraction(d) {
-    /*
-    Convert float to its approximate fractional representation. '''
-
-    This code was translated to JavaScript from the answers at
-    http://stackoverflow.com/questions/95727/how-to-convert-floats-to-human-\
-readable-fractions/681534#681534
-
-    For example:
-    >>> 3./5
-    0.59999999999999998
-
-    >>> rationalToFraction(3./5)
-    "3/5"
-
-    */
-    if (d === 0 || isNaN(d) || !isFinite(d)) {
-        return [0, 1];
-    }
-
-    let invert;
-    if (d > 1) {
-        invert = true;
-        d = 1 / d;
-    } else {
-        invert = false;
-    }
-
-    let df = 1.0;
-    let top = 1;
-    let iterations = 0;
-    const maxIterations = 10000;
-    let bot = 1;
-
-    while (Math.abs(df - d) > 0.00000001 && iterations < maxIterations) {
-        if (df < d) {
-            top += 1;
-        } else {
-            bot += 1;
-            top = Math.round(d * bot);
-        }
-
-        df = top / bot;
-        iterations++;
-    }
-
-    if (iterations === maxIterations) {
-        //console.warn("rationalToFraction: Reached iteration limit");
-        return [top, bot];
-    }
-
-    if (bot === 0 || top === 0) {
-        return [0, 1];
-    }
-
-    if (invert) {
-        return [bot, top];
-    } else {
-        return [top, bot];
-    }
-}
-
-/**
- * Calculates the greatest common divisor (GCD) of two numbers.
- * @param {number} a - The first number.
- * @param {number} b - The second number.
- * @returns {number} The GCD of the two numbers.
- */
-function GCD(a, b) {
-    a = Math.abs(a);
-    b = Math.abs(b);
-
-    while (b) {
-        const n = b;
-        b = a % b;
-        a = n;
-    }
-
-    return a;
-}
-
-/**
- * Converts a number to a mixed fraction string.
- * @param {number} d - The input number.
- * @returns {string} The mixed fraction string.
- */
-let mixedNumber = d => {
-    // Return number as a mixed fraction string, e.g., "2 1/4"
-
-    if (typeof d === "number") {
-        const floor = Math.floor(d);
-        if (d > floor) {
-            const obj = rationalToFraction(d - floor);
-            if (floor === 0) {
-                return obj[0] + "/" + obj[1];
-            } else {
-                if (obj[0] === 1 && obj[1] === 1) {
-                    return floor + 1;
-                } else {
-                    if (obj[1] > 99) {
-                        return d.toFixed(2);
-                    } else {
-                        return floor + " " + obj[0] + "/" + obj[1];
-                    }
-                }
-            }
-        } else {
-            return d.toString() + "/1";
-        }
-    } else {
-        return d;
-    }
-};
-
-/**
- * Calculates the least common denominator (LCD) of two numbers.
- * @param {number} a - The first number.
- * @param {number} b - The second number.
- * @returns {number} The LCD of the two numbers.
- */
-const LCD = (a, b) => {
-    return Math.abs((a * b) / GCD(a, b));
-};
-
-/**
- * Adds two rational numbers represented as arrays [numerator, denominator].
- *
- * This helper is used internally where rational arithmetic is required
- * to avoid floating-point precision issues (e.g., turtle singer logic).
- *
- * @param {Array} a - The first rational number.
- * @param {Array} b - The second rational number.
- * @returns {Array} The sum of the two rational numbers in the form [numerator, denominator].
- */
-let rationalSum = (a, b) => {
-    if (
-        !Array.isArray(a) ||
-        a.length < 2 ||
-        !Array.isArray(b) ||
-        b.length < 2 ||
-        typeof a[0] !== "number" ||
-        typeof a[1] !== "number" ||
-        typeof b[0] !== "number" ||
-        typeof b[1] !== "number" ||
-        a[1] === 0 ||
-        b[1] === 0
-    ) {
-        console.warn("Invalid input passed to rationalSum:", a, b);
-        return [0, 1];
-    }
-
-    // Make sure a and b components are integers.
-    let obja0, objb0, obja1, objb1;
-    if (Math.floor(a[0]) !== a[0]) {
-        obja0 = rationalToFraction(a[0]);
-    } else {
-        obja0 = [a[0], 1];
-    }
-
-    if (Math.floor(b[0]) !== b[0]) {
-        objb0 = rationalToFraction(b[0]);
-    } else {
-        objb0 = [b[0], 1];
-    }
-
-    if (Math.floor(a[1]) !== a[1]) {
-        obja1 = rationalToFraction(a[1]);
-    } else {
-        obja1 = [a[1], 1];
-    }
-
-    if (Math.floor(b[1]) !== b[1]) {
-        objb1 = rationalToFraction(b[1]);
-    } else {
-        objb1 = [b[1], 1];
-    }
-    // Use local variables to avoid mutating the caller's arrays
-    const a0 = obja0[0] * obja1[1];
-    const a1 = obja0[1] * obja1[0];
-    const b0 = objb0[0] * objb1[1];
-    const b1 = objb0[1] * objb1[0];
-
-    // Find the least common denomenator
-    const lcd = LCD(a1, b1);
-    return [(a0 * lcd) / a1 + (b0 * lcd) / b1, lcd];
-};
-
-/**
- * Finds the nearest beat for a given fraction.
- * @param {number} d - The numerator of the fraction.
- * @param {number} b - The denominator of the fraction.
- * @returns {Array} An array representing the nearest beat in the form [numerator, denominator].
- */
-let nearestBeat = (d, b) => {
-    // Find the closest beat for a given fraction.
-
-    let sum = 1 / (2 * b);
-    let count = 0;
-    const dd = d / 100;
-    while (dd > sum) {
-        sum += 1 / b;
-        count += 1;
-    }
-
-    return [count, b];
-};
-
-/**
- * Generates simple fractions based on a scale of 1-100.
- * @param {number} d - The input number.
- * @returns {Array} An array representing the fraction in the form [numerator, denominator].
- */
-let oneHundredToFraction = d => {
-    // Generate some simple fractions based on a scale of 1-100
-
-    if (d < 1) {
-        return [1, 64];
-    } else if (d > 99) {
-        return [1, 1];
-    }
-
-    switch (Math.floor(d)) {
-        case 1:
-            return [1, 64];
-        case 2:
-            return [1, 48];
-        case 3:
-        case 4:
-        case 5:
-            return [1, 32];
-        case 6:
-        case 7:
-        case 8:
-            return [1, 16];
-        case 9:
-        case 10:
-        case 11:
-            return [1, 12];
-        case 12:
-        case 13:
-        case 14:
-            return [1, 8];
-        case 15:
-        case 16:
-        case 17:
-            return [1, 6];
-        case 18:
-        case 19:
-            return [3, 16];
-        case 20:
-        case 21:
-        case 22:
-            return [1, 5];
-        case 23:
-        case 24:
-        case 25:
-        case 26:
-        case 27:
-        case 28:
-        case 29:
-            return [1, 4];
-        case 30:
-        case 31:
-            return [5, 16];
-        case 32:
-        case 33:
-        case 34:
-        case 35:
-            return [1, 3];
-        case 36:
-        case 37:
-        case 38:
-        case 39:
-            return [3, 8];
-        case 40:
-        case 41:
-            return [2, 5];
-        case 42:
-        case 43:
-        case 44:
-            return [7, 16];
-        case 45:
-        case 46:
-        case 47:
-            return [15, 32];
-        case 48:
-        case 49:
-        case 50:
-        case 51:
-        case 52:
-            return [1, 2];
-        case 53:
-        case 54:
-            return [17, 32];
-        case 56:
-        case 57:
-        case 58:
-            return [9, 16];
-        case 59:
-        case 60:
-        case 61:
-            return [3, 5];
-        case 62:
-        case 63:
-        case 64:
-        case 65:
-            return [5, 8];
-        case 66:
-        case 67:
-            return [2, 3];
-        case 68:
-        case 69:
-        case 70:
-            return [11, 16];
-        case 71:
-        case 72:
-        case 73:
-        case 74:
-            return [23, 32];
-        case 75:
-        case 76:
-        case 77:
-        case 78:
-        case 79:
-        case 80:
-            return [3, 4];
-        case 81:
-        case 82:
-            return [13, 16];
-        case 83:
-        case 84:
-        case 85:
-        case 86:
-            return [5, 6];
-        case 87:
-        case 88:
-        case 89:
-        case 90:
-            return [7, 8];
-        case 91:
-        case 92:
-            return [11, 12];
-        case 93:
-        case 94:
-        case 95:
-            return [15, 16];
-        case 96:
-        case 98:
-            return [31, 32];
-        case 99:
-            return [63, 64];
-        default:
-            return [d, 100];
-    }
-};
-
-/**
- * Converts RGB values to a hexadecimal color code.
- *
- * @param {number} r - Red value (0-255).
- * @param {number} g - Green value (0-255).
- * @param {number} b - Blue value (0-255).
- * @returns {string} Hexadecimal color code.
- */
-let rgbToHex = (r, g, b) => {
-    return "#" + ((1 << 24) + (r << 16) + (g << 8) + b).toString(16).slice(1);
-};
-
-/**
- * Converts a hexadecimal color code to RGB values.
- *
- * @param {string} hex - Hexadecimal color code.
- * @returns {Object} Object with RGB values {r, g, b} or null if invalid hex code.
- */
-let hexToRGB = hex => {
-    const result = /^#?([a-f\d]{2})([a-f\d]{2})([a-f\d]{2})$/i.exec(hex);
-    return result
-        ? {
-              r: parseInt(result[1], 16),
-              g: parseInt(result[2], 16),
-              b: parseInt(result[3], 16)
-          }
-        : null;
-};
-
-/**
- * Converts a hexcode to RGBA format.
- *
- * @param {number} hex - Hexadecimal color code.
- * @returns {string} RGBA color value with alpha set to 1.
- */
-let hex2rgb = hex => {
-    const bigint = parseInt(hex, 16);
-    const r = (bigint >> 16) & 255;
-    const g = (bigint >> 8) & 255;
-    const b = bigint & 255;
-
-    return "rgba(" + r + "," + g + "," + b + ",1)";
-};
-
-/**
  * Delays execution using a promise.
- *
- * @param {number} duration - Duration of the delay in milliseconds.
- * @returns {Promise} A promise that resolves after the specified duration.
+ * @requires browser (setTimeout)
  */
 let delayExecution = duration => {
     return new Promise(resolve => {
@@ -1960,8 +1286,7 @@ let delayExecution = duration => {
 
 /**
  * Closes all widgets in the window.
- *
- * @returns {void}
+ * @requires browser
  */
 function closeWidgets() {
     window.widgetWindows.hideAllWindows();
@@ -1969,9 +1294,7 @@ function closeWidgets() {
 
 /**
  * Closes a specific widget by its name.
- *
- * @param {string} name - The name of the widget to be closed.
- * @returns {void}
+ * @requires browser
  */
 let closeBlkWidgets = name => {
     const widgetTitle = document.getElementsByClassName("wftTitle");
@@ -1984,49 +1307,9 @@ let closeBlkWidgets = name => {
 };
 
 /**
- * Safely resolves a dot-notation string path to an object property globally.
- * @param {string} path - The dot-notation path (e.g., "MyClass.Model").
- * @returns {Object|undefined} The resolved object or undefined.
- */
-const resolveObject = path => {
-    if (!path || typeof path !== "string") return undefined;
-
-    // Support both browser and Node.js environments
-    const globalObj = typeof window !== "undefined" ? window : global;
-
-    try {
-        const result = path.split(".").reduce((obj, prop) => {
-            if (obj === null || obj === undefined) {
-                return undefined;
-            }
-            return obj[prop];
-        }, globalObj);
-
-        return result;
-    } catch (e) {
-        console.warn("Failed to resolve object path: " + path, e);
-        return undefined;
-    }
-};
-
-/**
  * Imports methods and variables of model and view objects to the controller object.
- *
- * @param {Object} obj - The component object (controller) to which members of its model and view are imported.
- * @param {string} className - Used for adding members of JS based MB API classes.
- * @param {*[]} modelArgs - Constructor arguments for the model.
- * @param {*[]} viewArgs - Constructor arguments for the view.
- * @returns {void}
  */
 let importMembers = (obj, className, modelArgs, viewArgs) => {
-    /**
-     * Adds methods and variables of one class to another class's instance.
-     *
-     * @param {Object} obj - Object of the component (controller).
-     * @param {Function} ctype - Static class type (model or view).
-     * @param {*[]} args - Array of constructor arguments.
-     * @returns {void}
-     */
     const addMembers = (obj, ctype, args) => {
         // If class type doesn't exist (no model class or no view class)
         if (ctype === undefined) {
@@ -2077,28 +1360,42 @@ let importMembers = (obj, className, modelArgs, viewArgs) => {
 };
 
 if (typeof module !== "undefined" && module.exports) {
+    const UtilsLogic = require("./utils-logic.js");
     module.exports = {
+        ...UtilsLogic,
         _,
-        last,
-        fileExt,
-        fileBasename,
-        toTitleCase,
-        safeSVG,
-        toFixed2,
-        mixedNumber,
-        nearestBeat,
-        oneHundredToFraction,
-        rationalToFraction,
-        rationalSum,
-        rgbToHex,
-        hexToRGB,
-        hex2rgb,
         format,
         delayExecution,
         closeWidgets,
         closeBlkWidgets,
-        resolveObject,
         importMembers,
-        escapeHTML
+        changeImage,
+        fnBrowserDetect,
+        canvasPixelRatio,
+        windowHeight,
+        windowWidth,
+        httpGet,
+        httpPost,
+        HttpRequest,
+        doBrowserCheck,
+        waitForReadiness,
+        docByClass,
+        docByTagName,
+        docById,
+        docByName,
+        docBySelector,
+        getTextWidth,
+        doSVG,
+        isSVGEmpty,
+        processPluginData,
+        processRawPluginData,
+        updatePluginObj,
+        preparePluginExports,
+        processMacroData,
+        prepareMacroExports,
+        doUseCamera,
+        doStopVideoCam,
+        hideDOMLabel,
+        displayMsg
     };
 }
